Data.ProtoLens.Default

Description

A compatibility layer for older code to create default protocol buffer messages.

In older versions of proto-lens, messages could be constructed with Data.Default.Class.def. However, for proto-lens >= 0.4, that is no longer the case and Data.ProtoLens.defMessage should be used instead.

This module provides a compatibility layer that may be used to upgrade older code without substantial code changes.

Documentation

def :: Message a => a

A message with all fields set to their default values.

For new code, prefer defMessage.

class Message msg

Every protocol buffer is an instance of Message. This class enables serialization by providing reflection of all of the fields that may be used by this type.

Minimal complete definition

messageName, packedMessageDescriptor, packedFileDescriptor, defMessage, fieldsByTag, unknownFields, parseMessage, buildMessage