Data.ProtoLens.Encoding.Bytes

Description

Utility functions for parsing and encoding individual types.

Running encodings

data Parser a

A monad for parsing an input buffer.

data Builder

Builders denote sequences of bytes. They are Monoids where mempty is the zero-length sequence and mappend is concatenation, which runs in O(1).

runParser :: Parser a -> ByteString -> Either String a

Evaluates a parser on the given input.

If the parser does not consume all of the input, the rest of the input is discarded and the parser still succeeds. Parsers may use atEnd to detect whether they are at the end of the input.

Values returned from actions in this monad will not hold onto the original ByteString, but rather make immutable copies of subsets of its bytes.

isolate :: Int -> Parser a -> Parser a

Run the given parsing action as if there are only len bytes remaining. That is, once len bytes have been consumed, atEnd will return True and other actions like getWord8 will act like there is no input remaining.

Fails the parse if given a negative length.

runBuilder :: Builder -> ByteString

Constructs a strict ByteString from the given Builder.

Bytestrings

getBytes :: Int -> Parser ByteString

Parse a sequence of zero or more bytes of the given length.

The new ByteString is an immutable copy of the bytes in the input and will be managed separately on the Haskell heap from the original input ByteString.

Fails the parse if given a negative length.

putBytes :: ByteString -> Builder

Emit a given ByteString.

Text

getText :: Int -> Parser Text

Integral types

getVarInt :: Parser Word64

getVarIntH :: Handle -> ExceptT String IO Word64

putVarInt :: Word64 -> Builder

getFixed32 :: Parser Word32

getFixed64 :: Parser Word64

putFixed32 :: Word32 -> Builder

putFixed64 :: Word64 -> Builder

Floating-point types

wordToFloat :: Word32 -> Float

wordToDouble :: Word64 -> Double

floatToWord :: Float -> Word32

doubleToWord :: Double -> Word64

Signed types

signedInt32ToWord :: Int32 -> Word32

wordToSignedInt32 :: Word32 -> Int32

signedInt64ToWord :: Int64 -> Word64

wordToSignedInt64 :: Word64 -> Int64

Other utilities

atEnd :: Parser Bool

Returns True if there is no more input left to consume.

runEither :: Either String a -> Parser a

(<?>) :: Parser a -> String -> Parser a

If the parser fails, prepend an error message.

foldMapBuilder :: Vector v a => (a -> Builder) -> v a -> Builder

Loop over the elements of a vector and concatenate the resulting Builders.

This function has been hand-tuned to perform better than a naive implementation using, e.g., Vector.foldr or a manual loop.