Data.ProtoLens.Encoding.Parser

Description

A custom parsing monad, optimized for speed.

Documentation

data Parser a

A monad for parsing an input buffer.

runParser :: Parser a -> ByteString -> Either String a

Evaluates a parser on the given input.

If the parser does not consume all of the input, the rest of the input is discarded and the parser still succeeds. Parsers may use atEnd to detect whether they are at the end of the input.

Values returned from actions in this monad will not hold onto the original ByteString, but rather make immutable copies of subsets of its bytes.

atEnd :: Parser Bool

Returns True if there is no more input left to consume.

isolate :: Int -> Parser a -> Parser a

Run the given parsing action as if there are only len bytes remaining. That is, once len bytes have been consumed, atEnd will return True and other actions like getWord8 will act like there is no input remaining.

Fails the parse if given a negative length.

getWord8 :: Parser Word8

Parse a one-byte word.

getWord32le :: Parser Word32

Parser a 4-byte word in little-endian order.

getBytes :: Int -> Parser ByteString

Parse a sequence of zero or more bytes of the given length.

The new ByteString is an immutable copy of the bytes in the input and will be managed separately on the Haskell heap from the original input ByteString.

Fails the parse if given a negative length.

getText :: Int -> Parser Text

(<?>) :: Parser a -> String -> Parser a

If the parser fails, prepend an error message.