Data.ProtoLens.Encoding.Wire

Description

Module defining the individual base wire types (e.g. VarInt, Fixed64).

They are used to represent the unknownFields within the proto message.

Upstream docs: https://developers.google.com/protocol-buffers/docs/encoding#structure

Documentation

newtype Tag

A tag that identifies a particular field of the message when converting to/from the wire format.

Constructors

Tag

Fields

unTag :: Int

data TaggedValue

A pair of an encoded field and a value.

Constructors

TaggedValue !Tag !WireValue

data WireValue

The encoding of some unknown field on the wire.

Constructors

VarInt !Word64

Fixed64 !Word64

Lengthy !ByteString

StartGroup

EndGroup

Fixed32 !Word32

type FieldSet = [TaggedValue]

splitTypeAndTag :: Word64 -> (Tag, Word8)

joinTypeAndTag :: Tag -> Word8 -> Word64

parseFieldSet :: Parser FieldSet

buildFieldSet :: FieldSet -> Builder

buildMessageSet :: FieldSet -> Builder

parseTaggedValueFromWire :: Word64 -> Parser TaggedValue

parseMessageSetTaggedValueFromWire :: Word64 -> Parser TaggedValue