Data.ProtoLens.Service.Types

Description

This module provides typeclasses for describing protobuf service metadata. It is intended to be used by library authors to generating bindings against proto services for specific RPC backends.

Documentation

class (KnownSymbol (ServiceName s), KnownSymbol (ServicePackage s), HasAllMethods s (ServiceMethods s)) => Service s where

Metadata describing a protobuf service. The s parameter is a phantom type generated by proto-lens.

The ServiceName and ServicePackage associated type can be used to generate RPC endpoint paths.

ServiceMethods is a promoted list containing every method defined on the service. As witnessed by the HasAllMethods superclass constraint here, this type can be used to discover every instance of HasMethod available for the service.

Associated Types

type ServiceName s :: Symbol

type ServicePackage s :: Symbol

type ServiceMethods s :: [Symbol]

Methods

packedServiceDescriptor :: Proxy s -> ByteString

class HasAllMethods s (ms :: [Symbol])

Reifies the fact that there is a HasMethod instance for every symbol claimed by the ServiceMethods associated type.

class (KnownSymbol m, KnownSymbol (MethodName s m), Service s, Message (MethodInput s m), Message (MethodOutput s m)) => HasMethodImpl s (m :: Symbol)

Metadata describing a service method. The MethodInput and MethodOutput type families correspond to the Messages generated by proto-lens for the RPC as described in the protobuf.

IsClientStreaming and IsServerStreaming can be used to dispatch on library code which wishes to provide different interfaces depending on the type of streaming of the method.

Library code should use HasMethod instead of this class directly whenever the constraint will be exposed to the end user. HasMethod provides substantially friendlier error messages when used incorrectly.

Associated Types

type MethodName s m :: Symbol

type MethodInput s m :: Type

type MethodOutput s m :: Type

type MethodStreamingType s m :: StreamingType

type HasMethod s m = (RequireHasMethod s m (ListContains m (ServiceMethods s)), HasMethodImpl s m)

Helper constraint that expands to a user-friendly error message when m isn't actually a method provided by service s.

data StreamingType

Data type to be used as a promoted type for MethodStreamingType.

Constructors

NonStreaming

ClientStreaming

ServerStreaming

BiDiStreaming