Data.ProtoLens.TextFormat

Description

Functions for converting protocol buffers to a human-readable text format.

Documentation

showMessage :: Message msg => msg -> String

Convert the given message into a human-readable String.

showMessageWithRegistry :: Message msg => Registry -> msg -> String

Convert the given message into a human-readable String, using the Registry to encode google.protobuf.Any values.

showMessageShort :: Message msg => msg -> String

Serializes a proto as a string on a single line. Useful for debugging and error messages like .DebugString() in other languages.

pprintMessage :: Message msg => msg -> Doc

Pretty-print the given message into a human-readable form.

pprintMessageWithRegistry :: Message msg => Registry -> msg -> Doc

Pretty-print the given message into human-readable form, using the given Registry to decode google.protobuf.Any values.

readMessage :: Message msg => Text -> Either String msg

Parse a Message from the human-readable protocol buffer text format.

readMessageWithRegistry :: Message msg => Registry -> Text -> Either String msg

Parse a Message from a human-readable protocol buffer text format, using the given Registry to decode Any fields

readMessageOrDie :: Message msg => Text -> msg

Parse a Message from the human-readable protocol buffer text format. Throws an error if the parse was not successful.